Revert "[FEAT] #39 - Title과 Divider ViewHolder/UiModel를 재사용할 수 있도록 각각 ui/model 모듈로 이동"
This reverts commit ccdad2f3b047fc8bf0d2953c8cf2a7b319fb7126.
