--- core/model/src/main/java/com/peonlee/model/divider/DividerUiModel.kt
@@ -1,8 +1,0 @@
-package com.peonlee.model.divider
-
-import com.peonlee.model.MainHomeListItem
-
-data class DividerUiModel(
-    override val id: Long,
-    override val viewType: Enum<*>
-) : MainHomeListItem
