--- core/ui/src/main/java/com/peonlee/core/ui/viewholder/divider/DividerViewHolder.kt
@@ -1,11 +1,0 @@
-package com.peonlee.core.ui.viewholder.divider
-
-import com.peonlee.core.ui.databinding.ListItemDividerBinding
-import com.peonlee.core.ui.viewholder.CommonViewHolder
-import com.peonlee.model.divider.DividerUiModel
-
-class DividerViewHolder(
-    binding: ListItemDividerBinding
-) : CommonViewHolder<DividerUiModel>(binding) {
-    override fun onBindView(item: DividerUiModel) {}
-}
