+++ feature/home/src/main/java/com/peonlee/home/HomeFragment.kt
@@ -4,12 +4,11 @@ import android.view.ViewGroup
 import com.peonlee.core.ui.base.BaseFragment
 import com.peonlee.home.adapter.HomeAdapter
 import com.peonlee.home.databinding.FragmentHomeBinding
+import com.peonlee.home.model.divider.DividerUiModel
 import com.peonlee.home.model.product.NEW_PRODUCTS
 import com.peonlee.home.model.product.POP_PRODUCTS
 import com.peonlee.home.model.review.RECENT_REVIEW
-import com.peonlee.model.MainHomeViewType
-import com.peonlee.model.divider.DividerUiModel
-import com.peonlee.model.title.TitleUiModel
+import com.peonlee.home.model.title.TitleUiModel
 
 class HomeFragment : BaseFragment<FragmentHomeBinding>() {
     override fun bindingFactory(parent: ViewGroup): FragmentHomeBinding {
@@ -30,23 +29,14 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>() {
 private val DUMMY = listOf(
     TitleUiModel(
         id = -1,
-        viewType = MainHomeViewType.TITLE,
         title = "주목할 신상"
     )
 ) + NEW_PRODUCTS + listOf(
     TitleUiModel(
         id = -2,
-        viewType = MainHomeViewType.TITLE,
         title = "꾸준한 인기상품이에요"
     )
 ) + POP_PRODUCTS + listOf(
-    DividerUiModel(
-        id = -3,
-        viewType = MainHomeViewType.DIVIDER
-    ),
-    TitleUiModel(
-        id = -4,
-        viewType = MainHomeViewType.TITLE,
-        title = "최근 리뷰"
-    )
+    DividerUiModel(id = -3),
+    TitleUiModel(id = -4, title = "최근 리뷰")
 ) + RECENT_REVIEW
