+++ feature/home/src/main/java/com/peonlee/home/model/divider/DividerUiModel.kt
@@ -1,0 +1,9 @@
+package com.peonlee.home.model.divider
+
+import com.peonlee.model.MainHomeListItem
+import com.peonlee.model.MainHomeViewType
+
+data class DividerUiModel(
+    override val id: Long,
+    override val viewType: Enum<MainHomeViewType> = MainHomeViewType.DIVIDER
+) : MainHomeListItem
