+++ feature/main/build.gradle.kts
@@ -10,6 +10,5 @@ android {
 dependencies {
     implementation(project(":feature:home"))
     implementation(project(":feature:explore"))
-    implementation(project(":feature:user"))
     implementation(libs.bundles.fragment)
 }
